+++ dev/benchmarks/test_apps/stocks/lib/stock_symbol_viewer.dart
@@ -139,6 +139,6 @@ class StockPage extends MaterialPage<void> {
     String symbol
   ) : super(
         key: const ValueKey<String>('stock'),
-        builder: (BuildContext context) => StockSymbolPage(symbol),
+        child: StockSymbolPage(symbol),
       );
 }
