+++ dev/benchmarks/test_apps/stocks/pubspec.yaml
@@ -2,7 +2,7 @@ name: stocks
 
 environment:
   # The pub client defaults to an <2.0.0 sdk constraint which we need to explicitly overwrite.
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
   flutter:
